requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
opencv-python-headless==4.8.0.74
pillow==9.5.0
numpy==1.23.5
scikit-learn>=1.0.0
joblib>=1.1.0 